+++ .gitignore
@@ -1,0 +1,12 @@
+# Node.js
+node_modules/*
+
+# Vite
+
+
+# IDE
+.vscode/
+.idea/
+
+# OS-generated files
+.DS_Store
