+++ src/components/Footer.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const Footer = () => {
+  return (<footer id="footer">
+    <div>
+      footer
+    </div>
+  </footer>);
+};
+
+export default Footer;
