+++ src/components/HeroSection.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import './../styles/Hero.css'
+const HeroSection = () => {
+  return (
+    <div id='hero-section' className='flex-perfect-center'>
+      <button id="shop-now-btn">Shop Now</button>
+    </div>
+  )
+}
+
+export default HeroSection
+
