+++ src/index.css
@@ -1,0 +1,55 @@
+@import url('https://fonts.googleapis.com/css2?family=Montserrat&display=swap');
+*{
+  font-family: 'Montserrat';
+  margin:0;
+  padding:0;
+  box-sizing: border-box;
+}
+#navbar{
+  width: 100vw;
+  position: fixed;
+  top:0;
+  left:0;
+  padding: 0.3rem;
+  background-color:rgba(14, 13, 13, 0.7);
+  backdrop-filter: blur(5px);
+  color:white;
+}
+#navbar .wrapper{
+  max-width:900px;
+  width: 100%;
+  margin: 0 0.8rem;
+  padding: 0 0.3rem;
+  justify-content: space-between;
+
+}
+#navbar h1{
+  text-align: center;
+}
+#navbar .menu-icon{
+  font-size: 2.5rem;
+}
+#navbar .shopping-bag{
+  font-size: 2rem;
+  background-color: transparent;
+  color: #fff;
+  padding: 0.3rem;
+  border: none;
+}
+
+
+/**-------globaluse case clases----*/
+.flex{
+  display:flex;
+}
+.justify-center{
+  justify-content: center;
+}
+.items-center{
+  align-items: center;
+}
+.flex-perfect-center{
+  display:flex;
+  justify-content: center;
+  align-items: center;
+}
