--- vite-project/src/components/Footer.jsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-const Footer = () => {
-  return <footer className="container">
-    hello from footer
-  </footer>;
-};
-
-export default Footer;
