--- vite-project/src/components/HeroSection.jsx
@@ -1,10 +1,0 @@
-import React from 'react'
-
-const HeroSection = () => {
-  return (
-    <div>HeroSection</div>
-  )
-}
-
-export default HeroSection
-
