--- vite-project/src/components/ProductsConatiner.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const ProductsConatiner = () => {
-  return (
-    <div>ProductsConatiner</div>
-  )
-}
-
-export default ProductsConatiner
